[Executor-types] move InMemoryStateCalculator to executor-types
