+++ execution/executor-types/Cargo.toml
@@ -13,6 +13,8 @@ edition = "2018"
 anyhow = "1.0.57"
 bcs = "0.1.3"
 itertools = "0.10.0"
+once_cell = "1.10.0"
+rayon = "1.5.2"
 serde = { version = "1.0.137", default-features = false }
 thiserror = "1.0.31"
