+++ execution/executor/src/components/mod.rs
@@ -7,4 +7,3 @@ pub mod apply_chunk_output;
 pub mod block_tree;
 pub mod chunk_commit_queue;
 pub mod chunk_output;
-pub mod in_memory_state_calculator;
