+++ state-sync/state-sync-v1/src/executor_proxy.rs
@@ -13,8 +13,7 @@ use aptos_types::{
     move_resource::MoveStorage, transaction::TransactionListWithProof,
 };
 use event_notifications::{EventNotificationSender, EventSubscriptionService};
-use executor::components::in_memory_state_calculator::IntoLedgerView;
-use executor_types::ChunkExecutorTrait;
+use executor_types::{in_memory_state_calculator::IntoLedgerView, ChunkExecutorTrait};
 use std::sync::Arc;
 use storage_interface::DbReader;
